# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WildTrack360 API",
    "version": "1.0.0",
    "description": "Internal API reference for WildTrack360. Generated from per-route Zod contracts."
  },
  "components": {
    "securitySchemes": {
      "clerkSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clerk session auth. In the browser the `__session` cookie is sent automatically; for the try-it client paste a session JWT as a bearer token."
      },
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "Shared secret for internal / scheduled (cron) endpoints."
      },
      "squareSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-square-hmacsha256-signature",
        "description": "Square webhook HMAC-SHA256 signature."
      }
    },
    "schemas": {
...
      "CarerInterest": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "clerkOrganizationId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONTACTED",
              "APPROVED",
              "DECLINED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "memberId": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "experience": {
            "type": [
              "string",
              "null"
            ]
          },
          "availability": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "clerkOrganizationId",
          "name",
          "email",
          "status",
          "createdAt",
          "updatedAt",
          "memberId",
          "phone",
          "experience",
          "availability",
          "message"
        ]
      },
      "CarerInterestOpen": {
        "type": "object",
        "properties": {
          "open": {
            "$ref": "#/components/schemas/CarerInterest"
          }
        },
        "required": [
          "open"
        ]
      },
      "CreatedId": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "CarerInterestSubmit": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "experience": {
            "type": "string"
          },
          "availability": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "HouseholdMembers": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "clerkUserId": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              },
              "required": [
                "id",
                "firstName",
                "lastName",
                "email",
                "clerkUserId"
              ]
            }
          }
        },
        "required": [
          "members"
        ]
      },
      "PortalOk": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "CheckoutResult": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string"
          },
          "squarePaymentId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "receiptNumber": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "paymentId",
          "squarePaymentId",
          "status",
          "receiptNumber"
        ]
      },
      "DonationCheckout": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "number"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "sourceId": {
            "type": "string",
            "minLength": 1
          },
          "verificationToken": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "amountCents",
          "sourceId"
        ]
      },
      "SubscriptionResult": {
        "type": "object",
        "properties": {
          "subscriptionId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "firstPaymentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "receiptNumber": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "subscriptionId",
          "status",
          "firstPaymentId",
          "receiptNumber"
        ]
      },
      "MembershipCheckout": {
        "type": "object",
        "properties": {
          "tierId": {
            "type": "string",
            "minLength": 1
          },
          "coverFees": {
            "type": "boolean"
          },
          "sourceId": {
            "type": "string",
            "minLength": 1
          },
          "verificationToken": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "tierId",
          "sourceId"
        ]
      },
      "RecurringDonationCheckout": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "number"
          },
          "interval": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "ANNUAL"
            ]
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "sourceId": {
            "type": "string",
            "minLength": 1
          },
          "verificationToken": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "amountCents",
          "interval",
          "sourceId"
        ]
      },
      "UpdateSubscriptionCard": {
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "string",
            "minLength": 1
          },
          "verificationToken": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "sourceId"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/portal/carer-interest": {
      "get": {
        "summary": "Get the current member open carer interest application",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Open carer interest or null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarerInterestOpen"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership found"
          }
        }
      },
      "post": {
        "summary": "Submit a carer interest application",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarerInterestSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created carer interest id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedId"
                }
              }
            }
          },
          "400": {
            "description": "Application already in progress or submission failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership found"
          }
        }
      }
    },
    "/api/portal/household": {
      "get": {
        "summary": "List household members",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Household member list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdMembers"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Secondary member or no active membership"
          },
          "404": {
            "description": "No membership found"
          }
        }
      },
      "post": {
        "summary": "Add a household member",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Created household member id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedId"
                }
              }
            }
          },
          "400": {
            "description": "Validation error from household lib"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Secondary member or no active membership"
          },
          "404": {
            "description": "No membership found"
          }
        }
      },
      "delete": {
        "summary": "Remove a household member",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalOk"
                }
              }
            }
          },
          "400": {
            "description": "id required or removal failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Secondary member or no active membership"
          },
          "404": {
            "description": "No membership found"
          }
        }
      }
    },
...
    "/api/portal/checkout/donation": {
      "post": {
        "summary": "Process a one-off donation payment",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonationCheckout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Donation payment result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or payment failure"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership or feature not enabled"
          }
        }
      }
    },
    "/api/portal/checkout/membership": {
      "post": {
        "summary": "Purchase a membership tier (creates a recurring subscription)",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCheckout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or payment failure"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership, feature not enabled, or tier not found"
          }
        }
      }
    },
    "/api/portal/checkout/recurring-donation": {
      "post": {
        "summary": "Start a recurring donation subscription",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringDonationCheckout"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResult"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or payment failure"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership or feature not enabled"
          }
        }
      }
    },
...
    "/api/portal/subscriptions/{id}/card": {
      "post": {
        "summary": "Update the card on file for a recurring subscription",
        "tags": [
          "Portal"
        ],
        "security": [
          {
            "clerkSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionCard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalOk"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or Square failure"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No membership, feature not enabled, or subscription not found"
          }
        }
      }
    }
  },
...
}